SignalProcessorHamasaki()をsignal_processor.pyに統合
